+++ src/lava/magma/runtime/message_infrastructure/test/CMakeLists.txt
@@ -12,11 +12,10 @@ include_directories(../message_infrastructure/csrc)
 
 add_executable(
   test_messaging_infrastructure
-  # cpptest/test_multiprocessing.cc
-  # cpptest/test_grpc_channel.cc
+  cpptest/test_multiprocessing.cc
   cpptest/test_grpc_delivery.cc
   # cpptest/test_channel.cc
-  # cpptest/test_shm_delivery.cc
+  cpptest/test_shm_delivery.cc
 )
 
 target_link_libraries(
